add: readme with current project state
This commit adds a README.md file that outlines a brief description of
what the project is and what the program currently supports. At current,
there is a limited CLI, which allows for a filename to be argued and the
disassembly flag to be argued; limited disassembly support, where only
register-to-register mov instructions are currently disassembled; and no
emulation support for the x86-16 architecture.